backend/requirements.txt
flask
flask-cors
opencv-python 
numpy
ultralytics
python-dotenv